Simulation to find optimal delta
This is a small Python simulation model that finds the optimal weighing of social information in a situation where individuals have to option to adopt a certain behaviour that has different payoff consequences for two types of individuals in a population. The difference in these payoff consequences is varied and also the frequency of both types in the population.